+++ pkg/driver/driver.go
@@ -18,35 +18,21 @@ package driver
 
 import (
 	"context"
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
-	"slices"
 	"time"
 
 	"github.com/google/cel-go/cel"
 	"github.com/google/dranet/pkg/apis"
-	"github.com/google/dranet/pkg/filter"
 	"github.com/google/dranet/pkg/inventory"
-	"github.com/google/dranet/pkg/names"
 
-	"github.com/Mellanox/rdmamap"
-	"github.com/containerd/nri/pkg/api"
 	"github.com/containerd/nri/pkg/stub"
 	"github.com/vishvananda/netlink"
-	"golang.org/x/sys/unix"
 
-	resourceapi "k8s.io/api/resource/v1beta1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/apimachinery/pkg/util/wait"
-	metav1apply "k8s.io/client-go/applyconfigurations/meta/v1"
-	resourceapply "k8s.io/client-go/applyconfigurations/resource/v1beta1"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/dynamic-resource-allocation/kubeletplugin"
-	"k8s.io/dynamic-resource-allocation/resourceslice"
 	"k8s.io/klog/v2"
 )
 
@@ -199,449 +185,6 @@ func (np *NetworkDriver) Stop() {
 	np.draPlugin.Stop()
 }
 
-func (np *NetworkDriver) Synchronize(_ context.Context, pods []*api.PodSandbox, containers []*api.Container) ([]*api.ContainerUpdate, error) {
-	klog.Infof("Synchronized state with the runtime (%d pods, %d containers)...",
-		len(pods), len(containers))
-
-	for _, pod := range pods {
-		klog.Infof("Synchronize Pod %s/%s UID %s", pod.Namespace, pod.Name, pod.Uid)
-		klog.V(2).Infof("pod %s/%s: namespace=%s ips=%v", pod.GetNamespace(), pod.GetName(), getNetworkNamespace(pod), pod.GetIps())
-		// get the pod network namespace
-		ns := getNetworkNamespace(pod)
-		// host network pods are skipped
-		if ns != "" {
-			// store the Pod metadata in the db
-			np.netdb.AddPodNetns(podKey(pod), ns)
-		}
-	}
-
-	return nil, nil
-}
-
 func (np *NetworkDriver) Shutdown(_ context.Context) {
 	klog.Info("Runtime shutting down...")
-}
-
-// CreateContainer handles container creation requests.
-func (np *NetworkDriver) CreateContainer(_ context.Context, pod *api.PodSandbox, ctr *api.Container) (*api.ContainerAdjustment, []*api.ContainerUpdate, error) {
-	klog.V(2).Infof("CreateContainer Pod %s/%s UID %s Container %s", pod.Namespace, pod.Name, pod.Uid, ctr.Name)
-	podConfig, ok := np.podConfigStore.GetPodConfigs(types.UID(pod.GetUid()))
-	if !ok {
-		return nil, nil, nil
-	}
-	// Containers only cares about the RDMA char devices
-	adjust := &api.ContainerAdjustment{}
-	for _, config := range podConfig {
-		for _, dev := range config.RDMADevice.DevChars {
-			adjust.AddDevice(&dev)
-		}
-	}
-	return adjust, nil, nil
-}
-
-func (np *NetworkDriver) RunPodSandbox(ctx context.Context, pod *api.PodSandbox) error {
-	klog.V(2).Infof("RunPodSandbox Pod %s/%s UID %s", pod.Namespace, pod.Name, pod.Uid)
-	start := time.Now()
-	defer func() {
-		klog.V(2).Infof("RunPodSandbox Pod %s/%s UID %s took %v", pod.Namespace, pod.Name, pod.Uid, time.Since(start))
-	}()
-	// get the devices associated to this Pod
-	podConfig, ok := np.podConfigStore.GetPodConfigs(types.UID(pod.GetUid()))
-	if !ok {
-		return nil
-	}
-	// get the pod network namespace
-	ns := getNetworkNamespace(pod)
-	// host network pods can not allocate network devices because it impact the host
-	if ns == "" {
-		return fmt.Errorf("RunPodSandbox pod %s/%s using host network can not claim host devices", pod.Namespace, pod.Name)
-	}
-	// store the Pod metadata in the db
-	np.netdb.AddPodNetns(podKey(pod), ns)
-
-	// Process the configurations of the ResourceClaim
-	errorList := []error{}
-	for deviceName, config := range podConfig {
-		klog.V(4).Infof("RunPodSandbox processing device: %s with config: %#v", deviceName, config)
-		resourceClaimStatus := resourceapply.ResourceClaimStatus()
-		// resourceClaim status for this specific device
-		resourceClaimStatusDevice := resourceapply.
-			AllocatedDeviceStatus().
-			WithDevice(deviceName).
-			WithDriver(np.driverName).
-			WithPool(np.nodeName)
-
-		ifName := names.GetOriginalName(deviceName)
-
-		klog.V(2).Infof("RunPodSandbox processing Network device: %s", ifName)
-		// TODO config options to rename the device and pass parameters
-		// use https://github.com/opencontainers/runtime-spec/pull/1271
-		networkData, err := nsAttachNetdev(ifName, ns, config.NetDevice)
-		if err != nil {
-			klog.Infof("RunPodSandbox error moving device %s to namespace %s: %v", deviceName, ns, err)
-			resourceClaimStatusDevice.WithConditions(
-				metav1apply.Condition().
-					WithType("Ready").
-					WithStatus(metav1.ConditionFalse).
-					WithReason("NetworkDeviceError").
-					WithMessage(err.Error()).
-					WithLastTransitionTime(metav1.Now()),
-			)
-			errorList = append(errorList, err)
-			continue
-		}
-
-		resourceClaimStatusDevice.WithConditions(
-			metav1apply.Condition().
-				WithType("Ready").
-				WithReason("NetworkDeviceReady").
-				WithStatus(metav1.ConditionTrue).
-				WithLastTransitionTime(metav1.Now()),
-		).WithNetworkData(resourceapply.NetworkDeviceData().
-			WithInterfaceName(networkData.InterfaceName).
-			WithHardwareAddress(networkData.HardwareAddress).
-			WithIPs(networkData.IPs...),
-		)
-		// configure routes
-		err = netnsRouting(ns, config.NetNamespaceRoutes)
-		if err != nil {
-			klog.Infof("RunPodSandbox error configuring device %s namespace %s routing: %v", deviceName, ns, err)
-			resourceClaimStatusDevice.WithConditions(
-				metav1apply.Condition().
-					WithType("NetworkReady").
-					WithStatus(metav1.ConditionFalse).
-					WithReason("NetworkReadyError").
-					WithMessage(err.Error()).
-					WithLastTransitionTime(metav1.Now()),
-			)
-		} else {
-			resourceClaimStatusDevice.WithConditions(
-				metav1apply.Condition().
-					WithType("NetworkReady").
-					WithStatus(metav1.ConditionTrue).
-					WithReason("NetworkReady").
-					WithLastTransitionTime(metav1.Now()),
-			)
-		}
-
-		// Move the RDMA device to the namespace if the host is in exclusive mode
-		if !np.rdmaSharedMode {
-			klog.V(2).Infof("RunPodSandbox processing RDMA device: %s", ifName)
-			err := nsAttachRdmadev(config.RDMADevice.LinkDev, ns)
-			if err != nil {
-				klog.Infof("RunPodSandbox error getting RDMA device %s to namespace %s: %v", deviceName, ns, err)
-				return err
-			}
-		}
-
-		resourceClaimStatus.WithDevices(resourceClaimStatusDevice)
-		resourceClaimApply := resourceapply.ResourceClaim(config.Claim.Name, config.Claim.Namespace).WithStatus(resourceClaimStatus)
-		// do not block the handler to update the status
-		go func() {
-			ctxStatus, cancel := context.WithTimeout(context.Background(), 3*time.Second)
-			defer cancel()
-			_, err = np.kubeClient.ResourceV1beta1().ResourceClaims(config.Claim.Namespace).ApplyStatus(ctxStatus,
-				resourceClaimApply,
-				metav1.ApplyOptions{FieldManager: np.driverName, Force: true},
-			)
-			if err != nil {
-				klog.Infof("failed to update status for claim %s/%s : %v", config.Claim.Namespace, config.Claim.Name, err)
-			} else {
-				klog.V(4).Infof("update status for claim %s/%s", config.Claim.Namespace, config.Claim.Name)
-			}
-		}()
-	}
-
-	return errors.Join(errorList...)
-}
-
-// StopPodSandbox tries to move back the devices to the rootnamespace but does not fail
-// to avoid disrupting the pod shutdown. The kernel will do the cleanup once the namespace
-// is deleted.
-func (np *NetworkDriver) StopPodSandbox(ctx context.Context, pod *api.PodSandbox) error {
-	klog.V(2).Infof("StopPodSandbox Pod %s/%s UID %s", pod.Namespace, pod.Name, pod.Uid)
-	start := time.Now()
-	defer func() {
-		np.netdb.RemovePodNetns(podKey(pod))
-		klog.V(2).Infof("StopPodSandbox Pod %s/%s UID %s took %v", pod.Namespace, pod.Name, pod.Uid, time.Since(start))
-	}()
-
-	// get the devices associated to this Pod
-	podConfig, ok := np.podConfigStore.GetPodConfigs(types.UID(pod.GetUid()))
-	if !ok {
-		return nil
-	}
-	// get the pod network namespace
-	ns := getNetworkNamespace(pod)
-	if ns == "" {
-		// some version of containerd does not send the network namespace information on this hook so
-		// we workaround it using the local copy we have in the db to associate interfaces with Pods via
-		// the network namespace id.
-		ns = np.netdb.GetPodNamespace(podKey(pod))
-		if ns == "" {
-			klog.Infof("StopPodSandbox pod %s/%s using host network ... skipping", pod.Namespace, pod.Name)
-			return nil
-		}
-	}
-
-	for deviceName, config := range podConfig {
-		ifName := names.GetOriginalName(deviceName)
-
-		if err := nsDetachNetdev(ns, config.NetDevice.Name, ifName); err != nil {
-			klog.Infof("fail to return network device %s : %v", deviceName, err)
-		}
-
-		if !np.rdmaSharedMode && config.RDMADevice.LinkDev != "" {
-			if err := nsDetachRdmadev(ns, config.RDMADevice.LinkDev); err != nil {
-				klog.Infof("fail to return rdma device %s : %v", deviceName, err)
-			}
-		}
-	}
-	return nil
-}
-
-func (np *NetworkDriver) RemovePodSandbox(_ context.Context, pod *api.PodSandbox) error {
-	klog.V(2).Infof("RemovePodSandbox Pod %s/%s UID %s", pod.Namespace, pod.Name, pod.Uid)
-	np.netdb.RemovePodNetns(podKey(pod))
-	return nil
-}
-
-func (np *NetworkDriver) PublishResources(ctx context.Context) {
-	klog.V(2).Infof("Publishing resources")
-	for {
-		select {
-		case devices := <-np.netdb.GetResources(ctx):
-			klog.V(4).Infof("Received %d devices", len(devices))
-			devices = filter.FilterDevices(np.celProgram, devices)
-			klog.V(4).Infof("After filtering %d devices", len(devices))
-			resources := resourceslice.DriverResources{
-				Pools: map[string]resourceslice.Pool{
-					np.nodeName: {Slices: []resourceslice.Slice{{Devices: devices}}}},
-			}
-			err := np.draPlugin.PublishResources(ctx, resources)
-			if err != nil {
-				klog.Error(err, "unexpected error trying to publish resources")
-			}
-		case <-ctx.Done():
-			klog.Error(ctx.Err(), "context canceled")
-			return
-		}
-		// poor man rate limit
-		time.Sleep(3 * time.Second)
-	}
-}
-
-func (np *NetworkDriver) PrepareResourceClaims(ctx context.Context, claims []*resourceapi.ResourceClaim) (map[types.UID]kubeletplugin.PrepareResult, error) {
-	klog.V(2).Infof("PrepareResourceClaims is called: number of claims: %d", len(claims))
-	if len(claims) == 0 {
-		return nil, nil
-	}
-	result := make(map[types.UID]kubeletplugin.PrepareResult)
-
-	for _, claim := range claims {
-		klog.V(2).Infof("NodePrepareResources: Claim Request %s/%s", claim.Namespace, claim.Name)
-		result[claim.UID] = np.prepareResourceClaim(ctx, claim)
-	}
-	return result, nil
-}
-
-// prepareResourceClaim gets all the configuration required to be applied at runtime and passes it downs to the handlers.
-// This happens in the kubelet so it can be a "slow" operation, so we can execute fast in RunPodsandbox, that happens in the
-// container runtime and has strong expectactions to be executed fast (default hook timeout is 2 seconds).
-func (np *NetworkDriver) prepareResourceClaim(ctx context.Context, claim *resourceapi.ResourceClaim) kubeletplugin.PrepareResult {
-	// TODO: shared devices may allocate the same device to multiple pods, i.e. macvlan, ipvlan, ...
-	podUIDs := []types.UID{}
-	for _, reserved := range claim.Status.ReservedFor {
-		if reserved.Resource != "pods" || reserved.APIGroup != "" {
-			klog.Infof("Driver only supports Pods, unsupported reference %#v", reserved)
-			continue
-		}
-		podUIDs = append(podUIDs, reserved.UID)
-	}
-	if len(podUIDs) == 0 {
-		klog.Infof("no pods allocated to claim %s/%s", claim.Namespace, claim.Name)
-		return kubeletplugin.PrepareResult{}
-	}
-
-	nlHandle, err := netlink.NewHandle()
-	if err != nil {
-		return kubeletplugin.PrepareResult{
-			Err: fmt.Errorf("error creating netlink handle %v", err),
-		}
-	}
-
-	var errorList []error
-	charDevices := sets.New[string]()
-	for _, result := range claim.Status.Allocation.Devices.Results {
-		requestName := result.Request
-		netconf := apis.NetworkConfig{}
-		for _, config := range claim.Status.Allocation.Devices.Config {
-			// Check there is a config associated to this device
-			if config.Opaque == nil ||
-				config.Opaque.Driver != np.driverName ||
-				len(config.Requests) > 0 && !slices.Contains(config.Requests, requestName) {
-				continue
-			}
-			// Check if there is a custom configuration
-			conf, err := apis.ValidateConfig(&config.Opaque.Parameters)
-			if err != nil {
-				errorList = append(errorList, err)
-				continue
-			}
-			// TODO: define a strategy for multiple configs
-			if conf != nil {
-				netconf = *conf
-				break
-			}
-		}
-		klog.V(4).Infof("PrepareResourceClaim %s/%s final Configuration %#v", claim.Namespace, claim.Name, netconf)
-		podCfg := PodConfig{
-			Claim: types.NamespacedName{
-				Namespace: claim.Namespace,
-				Name:      claim.Name,
-			},
-			NetDevice:          netconf.Interface,
-			NetNamespaceRoutes: netconf.Routes,
-		}
-		ifName := names.GetOriginalName(result.Device)
-		// Get Network configuration and merge it
-		link, err := nlHandle.LinkByName(ifName)
-		if err != nil {
-			errorList = append(errorList, fmt.Errorf("fail to get network interface %s", ifName))
-			continue
-		}
-
-		// If there is no custom addresses then use the existing ones
-		if len(podCfg.NetDevice.Addresses) == 0 {
-			// get the existing IP addresses
-			nlAddresses, err := nlHandle.AddrList(link, netlink.FAMILY_ALL)
-			if err != nil && !errors.Is(err, netlink.ErrDumpInterrupted) {
-				klog.Infof("fail to get ip addresses for interface %s : %v", ifName, err)
-			} else {
-				for _, address := range nlAddresses {
-					// Only move IP addresses with global scope because those are not host-specific, auto-configured,
-					// or have limited network scope, making them unsuitable inside the container namespace.
-					// Ref: https://www.ietf.org/rfc/rfc3549.txt
-					if address.Scope != unix.RT_SCOPE_UNIVERSE {
-						continue
-					}
-					podCfg.NetDevice.Addresses = append(podCfg.NetDevice.Addresses, address.IPNet.String())
-				}
-			}
-		}
-
-		// If there are no addresses configured on the interface and the user is not setting them
-		// this may be an interface that uses DHCP, so we bring it up if necessary and do a DHCP
-		// request to gather the network parameters (IPs and Routes) ... but we DO NOT apply them
-		// in the root namespace
-		if len(podCfg.NetDevice.Addresses) == 0 {
-			klog.V(2).Infof("trying to get network configuration via DHCP")
-			ip, routes, err := getDHCP(ifName)
-			if err != nil {
-				klog.Infof("fail to get configuration via DHCP: %v", err)
-			} else {
-				podCfg.NetDevice.Addresses = []string{ip}
-				podCfg.NetNamespaceRoutes = append(podCfg.NetNamespaceRoutes, routes...)
-			}
-		}
-
-		// Obtain the routes associated to the interface
-		// TODO: only considers outgoing traffic
-		filter := &netlink.Route{
-			LinkIndex: link.Attrs().Index,
-		}
-		routes, err := nlHandle.RouteListFiltered(netlink.FAMILY_ALL, filter, netlink.RT_FILTER_OIF)
-		if err != nil {
-			klog.Infof("fail to get ip routes for interface %s : %v", ifName, err)
-		}
-		for _, route := range routes {
-			routeCfg := apis.RouteConfig{}
-
-			if route.Gw == nil || route.Dst == nil {
-				continue
-			}
-			if !route.Dst.IP.IsGlobalUnicast() {
-				continue
-			}
-			routeCfg.Gateway = route.Gw.String()
-			routeCfg.Destination = route.Dst.String()
-			if route.Src != nil {
-				routeCfg.Source = route.Src.String()
-			}
-			podCfg.NetNamespaceRoutes = append(podCfg.NetNamespaceRoutes, routeCfg)
-		}
-
-		// Get RDMA configuration: link and char devices
-		if rdmaDev, _ := rdmamap.GetRdmaDeviceForNetdevice(ifName); rdmaDev != "" {
-			klog.V(2).Infof("RunPodSandbox processing RDMA device: %s", rdmaDev)
-			podCfg.RDMADevice.LinkDev = rdmaDev
-			// Obtain the char devices associated to the rdma device
-			charDevices.Insert(rdmaCmPath)
-			charDevices.Insert(rdmamap.GetRdmaCharDevices(rdmaDev)...)
-			for _, devpath := range charDevices.UnsortedList() {
-				dev, err := GetDeviceInfo(devpath)
-				if err != nil {
-					klog.Infof("fail to get device info for %s : %v", devpath, err)
-				} else {
-					podCfg.RDMADevice.DevChars = append(podCfg.RDMADevice.DevChars, dev)
-				}
-			}
-		}
-
-		device := kubeletplugin.Device{
-			Requests:   []string{result.Request},
-			PoolName:   result.Pool,
-			DeviceName: result.Device,
-		}
-		// TODO: support for multiple pods sharing the same device
-		// we'll create the subinterface here
-		for _, uid := range podUIDs {
-			np.podConfigStore.Set(uid, device.DeviceName, podCfg)
-		}
-		klog.V(4).Infof("Claim Resources for pods %v : %#v", podUIDs, podCfg)
-	}
-
-	if len(errorList) > 0 {
-		klog.Infof("claim %s contain errors: %v", claim.UID, errors.Join(errorList...))
-		return kubeletplugin.PrepareResult{
-			Err: fmt.Errorf("claim %s contain errors: %w", claim.UID, errors.Join(errorList...)),
-		}
-	}
-	return kubeletplugin.PrepareResult{}
-}
-
-func (np *NetworkDriver) UnprepareResourceClaims(ctx context.Context, claims []kubeletplugin.NamespacedObject) (map[types.UID]error, error) {
-	klog.V(2).Infof("UnprepareResourceClaims is called: number of claims: %d", len(claims))
-	if len(claims) == 0 {
-		return nil, nil
-	}
-
-	result := make(map[types.UID]error)
-	for _, claim := range claims {
-		err := np.unprepareResourceClaim(ctx, claim)
-		result[claim.UID] = err
-		if err != nil {
-			klog.Infof("error unpreparing ressources for claim %s/%s : %v", claim.Namespace, claim.Name, err)
-		}
-	}
-	return result, nil
-}
-
-func (np *NetworkDriver) unprepareResourceClaim(_ context.Context, claim kubeletplugin.NamespacedObject) error {
-	np.podConfigStore.DeleteClaim(claim.NamespacedName)
-	return nil
-}
-
-func getNetworkNamespace(pod *api.PodSandbox) string {
-	// get the pod network namespace
-	for _, namespace := range pod.Linux.GetNamespaces() {
-		if namespace.Type == "network" {
-			return namespace.Path
-		}
-	}
-	return ""
-}
-
-func podKey(pod *api.PodSandbox) string {
-	return fmt.Sprintf("%s/%s", pod.GetNamespace(), pod.GetName())
 }
